requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==4.2
django-cors-headers==4.3.1
django-filter==24.2
django-semantic-admin==0.5.0
django-semantic-forms==0.1.5
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
et-xmlfile==1.1.0
idna==3.7
Markdown==3.6
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
